refactoring :: Impl
