+++ app/src/main/java/com/hseun/lendy_v2/splash/viewmodel/SplashViewModel.kt
@@ -5,12 +5,12 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.hseun.lendy_v2.splash.repository.SplashRepository
+import com.hseun.lendy_v2.splash.repository.SplashRepositoryImpl
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 class SplashViewModel @Inject constructor(
-    private val repository: SplashRepository
+    private val repository: SplashRepositoryImpl
 ) : ViewModel() {
     var isAutoLoginCheck by mutableStateOf(false)
         private set
